Initial Commit
I want to create a replacement for my current python script for
downloading and archiving podcasts. The limitations of the current
script are that it's brittle, it can't handle all RSS feeds, and it's
logs leave a lot to be desired. I'm hoping a C# rewrite will can resolve
these issues. I am choosing C# so this can also be a practice project
for my day job.