Adds ability to read in configuration file config.xml
This config.xml file is how the user will specify what podcasts they
want to download. Right now we just look in the working directory for
the file, but I want future enhancements to look in several locations.
If a file isn't found, I want the program to prompt the user for some
information and create one. I want there to be commands in the cli that
allow you to modify the config file, so the user doesn't have to know it
exists if they don't want to.